pyproject.toml
[project]
name = "stock-app-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "celery>=5.5.3",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.4",
    "django-allauth>=65.9.0",
    "django-antispam>=1.0.3",
    "django-celery-beat>=2.8.1",
    "django-debug-toolbar>=5.2.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-htmx>=1.23.2",
    "django-redis>=6.0.0",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "eventlet>=0.40.1",
    "fcm-django>=2.2.1",
    "fear-and-greed>=0.4",
    "finviz>=1.4.6",
    "finvizfinance>=1.1.1",
    "google-genai>=1.39.1",
    "gunicorn>=23.0.0",
    "gymnasium>=1.2.0",
    "html5lib>=1.1",
    "locust>=2.37.14",
    "lxml>=6.0.0",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.0",
    "pip>=25.1.1",
    "psycopg2-binary>=2.9.10",
    "pyprof2calltree>=1.4.5",
    "python-decouple>=3.8",
    "redis>=6.2.0",
    "scikit-learn>=1.7.1",
    "setuptools>=80.9.0",
    "stock-indicators>=1.3.5",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "werkzeug>=3.1.3",
    "yfinance>=0.2.65",
]
